+++ CMakeLists.txt
@@ -261,6 +261,7 @@ add_library(
   src/s2geography/build.cc
   src/s2geography/coverings.cc
   src/s2geography/distance.cc
+  src/s2geography/geoarrow-geography.cc
   src/s2geography/geoarrow.cc
   src/s2geography/geography.cc
   src/s2geography/linear-referencing.cc
@@ -268,7 +269,6 @@ add_library(
   src/s2geography/op/point.cc
   src/s2geography/predicates.cc
   src/s2geography/projections.cc
-  src/s2geography/wkb-geography.cc
   src/s2geography/wkb.cc
   src/s2geography/wkt-reader.cc
   src/s2geography/wkt-writer.cc
@@ -355,7 +355,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   add_executable(predicates_test src/s2geography/predicates_test.cc)
   add_executable(wkt_writer_test src/s2geography/wkt-writer_test.cc)
   add_executable(wkb_test src/s2geography/wkb_test.cc)
-  add_executable(wkb_geography_test src/s2geography/wkb-geography_test.cc)
+  add_executable(geoarrow_geography_test src/s2geography/geoarrow-geography_test.cc)
 
   if(S2GEOGRAPHY_CODE_COVERAGE)
     target_compile_options(coverage_config INTERFACE -O0 -g --coverage)
@@ -389,7 +389,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
     GTest::gtest_main GTest::gmock)
   target_link_libraries(wkt_writer_test s2geography GTest::gtest_main)
   target_link_libraries(wkb_test s2geography GTest::gtest_main)
-  target_link_libraries(wkb_geography_test s2geography GTest::gtest_main)
+  target_link_libraries(geoarrow_geography_test s2geography GTest::gtest_main)
 
   target_include_directories(sedona_udf_internal_test PRIVATE src/vendored)
   target_include_directories(accessors_geog_test PRIVATE src/vendored)
@@ -398,7 +398,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   target_include_directories(geoarrow_test PRIVATE src/vendored)
   target_include_directories(linear_referencing_test PRIVATE src/vendored)
   target_include_directories(predicates_test PRIVATE src/vendored)
-  target_include_directories(wkb_geography_test PRIVATE src/vendored)
+  target_include_directories(geoarrow_geography_test PRIVATE src/vendored)
 
   include(GoogleTest)
   gtest_discover_tests(sedona_udf_internal_test)
@@ -412,7 +412,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   gtest_discover_tests(geography_test)
   gtest_discover_tests(wkt_writer_test)
   gtest_discover_tests(wkb_test)
-  gtest_discover_tests(wkb_geography_test)
+  gtest_discover_tests(geoarrow_geography_test)
 endif()
 
 if(S2GEOGRAPHY_BUILD_EXAMPLES)
